+++ src/Monify.Tests/Snippets/Declarations/Classes/Nested.InRecordStruct.Expected.NonNullable.cs
@@ -1031,6 +1031,11 @@ internal static partial class Nested
                             {
                                 public int CompareTo(object value)
                                 {
+                                    if (value is Inner)
+                                    {
+                                        value = ((Inner)value)._value;
+                                    }
+
                                     return _value.CompareTo(value);
                                 }
                             }