+++ src/Monify.Tests/Snippets/Declarations/Classes/Nested.MultiLevel.Expected.Nullable.cs
@@ -1369,6 +1369,11 @@ internal static partial class Nested
                                 {
                                     public int CompareTo(object value)
                                     {
+                                        if (value is InlineStyle)
+                                        {
+                                            value = ((InlineStyle)value)._value;
+                                        }
+
                                         return _value.CompareTo(value);
                                     }
                                 }